+++ sphinx/source/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 ---------
 * Added the ``-p`` option to :ref:`skool2html.py` (to show the path to the
   skoolkit package directory)
+* Added the ``-T`` option to :ref:`skool2html.py` (to specify a CSS theme)
 * Fixed the appearance of transparent table cells in HTML output 
 * Added `jet_set_willy-dark.css` (to complete the 'dark' theme for that
   disassembly)